Cumulative update

+++ ConsoleRenderer/Extensions/WebcamInConsole/ColorPair.cs
@@ -9,8 +9,8 @@ namespace NostalgiaEngine.Extensions
 {
 	class ColorPair
 	{
-		public int CI0;
-		public int CI1;
+		public byte CI0;
+		public byte CI1;
 		public NEColorPalette Palette;
 
 		NEVector4 col0;
@@ -20,7 +20,7 @@ namespace NostalgiaEngine.Extensions
 		public float ColDistance;
 
 		public float ColDistanceInv;
-		public ColorPair(int ci0, int ci1, NEColorPalette palette)
+		public ColorPair(byte ci0, byte ci1, NEColorPalette palette)
 		{
 			CI0 = ci0;
 			CI1 = ci1;
@@ -53,7 +53,7 @@ namespace NostalgiaEngine.Extensions
 		public static ColorPair[] GenerateColorPairs(NEColorPalette palette)
 		{
 			List<ColorPair> pairs = new List<ColorPair>();
-			for (int i = 1; i < 16; ++i)
+			for (byte i = 1; i < 16; ++i)
 			{
 				pairs.Add(new ColorPair(0, i, palette));
 			}
